requirements.txt
Django>=4.2.7,<5.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.1.0
whitenoise>=6.6.0
gunicorn>=21.2.0
django-storages[boto3]>=1.14.0
boto3>=1.34.0
